+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Last Changes
 
+
+## v1.0.0-alpha.0
+
 - [#14](https://github.com/LaxarJS/ax-details-layer-widget/issues/14): added travis CI
 - [#13](https://github.com/LaxarJS/ax-details-layer-widget/issues/13): fixed close animation appearance
 - [#10](https://github.com/LaxarJS/ax-details-layer-widget/issues/10): allow to reflect and control visibility state in URL